fix timezones in darcs-fast-export, take 2
The previous patch had a flaw; it assumed that darcs was storing the committer's timezone.
Instead, darcs always stores UTC timestamps in an ISO 8601 format.
Tools like "darcs changes" convert this into the user's local time as a convenience.

I couldn't find an authoritative spec, but here are some relevant references.
http://wiki.darcs.net/NamedPatch
http://search.cpan.org/~david/Darcs-Inventory-1.4/lib/Darcs/Inventory/Patch.pm
http://bugs.darcs.net/issue140

To resolve the issue, this patch always reports that the timezone is UTC.