Allow different formats for tag completion.
The tests now use the dirstate-tags format, as does the
bzrlib.tests.test_tag module.  As "bzr tags" will print a warning about a
deprecated format for this, we should discard its error output.

That seems a good idea for production environments as well, as outdated
formats might be quite frequent, and having completion print error messages
for all of these is certainly bad style.