Add --directory option to whoami.
Like for the cat command for example, if a --directory option is given, it
has to denote the root of a branch and may be a URL to a remote branch.  If
the named path or URL doesn't denote a branch, an error is reported, instead
of a silent fallback to the global configuration.