Avoid duplicate file ids in the import code with renames.
Don't crash when doing merge-upstream with a branch that does a rename
and then ships another file with the old path in the tarball that isn't
in the branch.

We would assign the same id to both paths, which obviously doesn't work.

Instead we now do a first pass to determine which file ids in the tree
will be assigned to different paths, and then avoid re-using that file
id with the old path.