Split file_ids_from in to one we should take renames from, and others.
Having multiple trees that we may follow renames from led to lots of
bugs, and turned out to be unnecessary.

Split file_ids_from to give a single tree that will be authoritative,
and a list of others that we will take file ids from, but only if we would
otherwise generate one.