Implement the 'brz config' command. Read-only.
* bzrlib/tests/blackbox/test_config.py:
Add some test for 'bzr config'.

* bzrlib/config.py:
(BranchConfig.get_options_matching_regexp): Let's stop to pretend
we can get a file name if the implementation doesn't provide it,
the known cases are bBranchConfig and RemoteBranchConfig that will
be modified anyway.
(cmd_config): A simple command to display options known for a
given directory.