Disabling lsprof shows something interesting.
It turns out that the bulk of the slowdown when searching for more revisions
was *specifically* the difference_update calls. As a table:

time  
25.2s short_search_gpm 
22.2s long_search_gpm.difference
34.2s long_search_gpm.difference_update

So I thought zlib.compressobj or bz2.compress was taking longer, but really
it was just that our 'queried_revs' count was getting so big it was making
difference_update exceedingly slow.
I'm very happy that the longer searches are now even faster over the loopback
than doing shorter prefetches. It makes a lot of sense, but it is nice to
see.