Catch one more detail in #842575 test case.
I found out that recreating the file will list it as added instead of
missing, despite the ignore.  So I've rearranged the tests so that the
ignore and the --take-other (which preserves the base name) are now in the
same test, and ensure that adding the file again will cause no output at
all.