Introduce an option "--omit-merges" for "bzr log".
This will omit the merging commits themselves, i.e. commits with more than a
single parent.  The idea is that these commits tend to describe the
mechanics of code development, but are at best only summaries of the work
done in sidelines.  So omitting them but including the sidelines will print
all the commits that actually contributed features to the code, without too
much noise and duplicate information.  This is particularly useful when
generating change logs for websites or tarballs.