Change the ordering a bit.
Wait for a 3rd connection to respond, which means we know we've:
a) created the connection, and polled for the old one being dead
b) fully added the third connection.

Another fix is to allow the shutdown code to handle when the
server already noticed that it needs to shut down, without
another connection attempt.