Bug #885690, Branch.revision_history() has been deprecated in bzr 2.5.
Many of the call sites actually only cared about 'last_revision()' so use that.
A few of them used revision_history() to get the tip revision, and the revision
prior to that, etc. So switch to the code from 'test_fetch'. I'm not
importing it directly, because that certainly isn't a stable location, and
would need a compatibility workaround for bzr-2.4, etc.