One of the discoveries from fixing bug #1046697.
A lightweight checkout of a remote location sprouting to a new location was not
properly passing along the newly created and initialized repository object to
'initialize' the new branch. This is because _vfs_initialize was ignoring the
parameter. So now we pass it along.
The test that failed was per_workingtree.test_get_parent_ids.TestGetParentIds.test_get_parent_ids.