Remove replacement of re.compile for lazy_regexp
On a typical invocation less than 10 regexp compilations are avoided
and most of those are in configobj. This is less than a millisecond
saving in general, which makes the complexity and other potential
issues with monkey patching probably not worth while.

Leave the LazyRegexp proxy object around which is used in various
places and generally harmless when expected.