Switch pypi package: gpg to pygpgme
pypi package gpg hasn't been updated in a while (latest, v1.10 was released on 2018-02-15).
REF: https://pypi.org/project/gpg/
REF: https://github.com/gpg/gpgme/releases/tag/gpgme-1.18.0

It has a hardcoded requirement of using the binary gpgme-config.
This binary was removed upstream in gpgme1.0 with v1.18.0-2
It was removed gpgme-config as it uses gpg-error-config
REF: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1023599

This means trying to install on or higher than Debian "Bookworm" 12 or Ubuntu "Noble Numbat" 24.04, will fail (even if the libraries/packages are installed).
REF: https://tracker.debian.org/pkg/libgpgme-dev
REF: https://packages.ubuntu.com/libgpgme-dev