launchpad: Add warning about Bazaar code hosting phase-out
When users access Bazaar branches on Launchpad (via lp: URLs or when
publishing branches), display a warning that Launchpad is phasing out
Bazaar code hosting. The warning includes a link to the announcement
blog post and can be suppressed by adding 'launchpad_bazaar_deprecation'
to the suppress_warnings configuration option.